refactor(cli): Add stryker as binary (#123)
Refactor Stryker class to not include the cli. Also add stryker as bin so it can be used directly from an npm script (i.e. "stryker": "stryker --configFile stryker.conf.js")
